+++ tests/unit/utility.hpp
@@ -7,7 +7,6 @@ copyright       MIT - Copyright (c) 2022 Oliver Blaser
 #ifndef TEST_OMW_UTIL_H
 #define TEST_OMW_UTIL_H
 
-#include <iostream>
 #include <string>
 #include <vector>
 
@@ -1492,81 +1491,6 @@ TEST_CASE("utility.h vectorContains()")
     const std::vector<omw::string> vString = { "pack", "my", "box", "with", "five", "dozen", "liquor", "jugs" };
     CHECK(omw::vectorContains(vString, omw::string("five")));
     CHECK_FALSE(omw::vectorContains(vString, omw::string("boat")));
-}
-
-
-
-//#if (defined(OMW_DEBUG) && 1) // || 1
-//#define UTILITY_TEST_VECRESERVEMORE_COUTEN 1
-//#endif
-//
-//TEST_CASE("utility.h vectorReserveMore()")
-//{
-//    std::vector<int> v;
-//
-//#ifdef UTILITY_TEST_VECRESERVEMORE_COUTEN
-//    std::cout << v.capacity() << std::endl;
-//#endif
-//    omw::vectorReserveMore(v, 5);
-//#ifdef UTILITY_TEST_VECRESERVEMORE_COUTEN
-//    std::cout << v.capacity() << std::endl;
-//#endif
-//    CHECK(v.capacity() >= (v.size() + 5));
-//
-//
-//    v.assign(50, -1);
-//#ifdef UTILITY_TEST_VECRESERVEMORE_COUTEN
-//    std::cout << v.capacity() << std::endl;
-//#endif
-//    omw::vectorReserveMore(v, 20);
-//#ifdef UTILITY_TEST_VECRESERVEMORE_COUTEN
-//    std::cout << v.capacity() << std::endl;
-//#endif
-//    CHECK(v.capacity() >= (v.size() + 20));
-//
-//
-//    v.shrink_to_fit();
-//#ifdef UTILITY_TEST_VECRESERVEMORE_COUTEN
-//    std::cout << v.capacity() << std::endl;
-//#endif
-//    omw::vectorReserveMore(v, 3);
-//#ifdef UTILITY_TEST_VECRESERVEMORE_COUTEN
-//    std::cout << v.capacity() << std::endl;
-//#endif
-//    CHECK(v.capacity() >= (v.size() + 3));
-//
-//
-//    for (size_t i = 0; i < 100; ++i) v.push_back(3);
-//#ifdef UTILITY_TEST_VECRESERVEMORE_COUTEN
-//    std::cout << v.capacity() << std::endl;
-//#endif
-//    omw::vectorReserveMore(v, 3);
-//#ifdef UTILITY_TEST_VECRESERVEMORE_COUTEN
-//    std::cout << v.capacity() << std::endl;
-//#endif
-//    CHECK(v.capacity() >= (v.size() + 3));
-//}
-
-
-
-TEST_CASE("utility.h convertByteVector()")
-{
-    const std::string initialString = "The quick brown fox jumps over the lazy dog.";
-    const std::vector<char> initial(initialString.c_str(), initialString.c_str() + initialString.length());
-    const std::vector<uint8_t> expectedResult =
-    {
-        0x54, 0x68, 0x65, 0x20, 0x71, 0x75, 0x69, 0x63, 0x6b, 0x20, 0x62, 0x72, 0x6f, 0x77, 0x6e, 0x20,
-        0x66, 0x6f, 0x78, 0x20, 0x6a, 0x75, 0x6d, 0x70, 0x73, 0x20, 0x6f, 0x76, 0x65, 0x72, 0x20, 0x74,
-        0x68, 0x65, 0x20, 0x6c, 0x61, 0x7a, 0x79, 0x20, 0x64, 0x6f, 0x67, 0x2e
-    };
-
-    const std::vector<uint8_t> r = omw::convertByteVector(initial);
-
-    CHECK(r == expectedResult);
-
-    const std::vector<char> r2 = omw::convertByteVector(r);
-
-    CHECK(r2 == initial);
 }